Move BlockStorage and PhantomBlock to block package
