+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/addon/PylonAddonListener.kt
@@ -1,7 +1,7 @@
 package io.github.pylonmc.pylon.core.addon
 
 import io.github.pylonmc.pylon.core.entity.EntityStorage
-import io.github.pylonmc.pylon.core.persistence.blockstorage.BlockStorage
+import io.github.pylonmc.pylon.core.block.BlockStorage
 import io.github.pylonmc.pylon.core.registry.PylonRegistry
 import org.bukkit.event.EventHandler
 import org.bukkit.event.Listener
