+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/block/MultiblockCache.kt
@@ -5,7 +5,6 @@ import io.github.pylonmc.pylon.core.event.PylonBlockBreakEvent
 import io.github.pylonmc.pylon.core.event.PylonBlockPlaceEvent
 import io.github.pylonmc.pylon.core.event.PylonChunkBlocksLoadEvent
 import io.github.pylonmc.pylon.core.event.PylonChunkBlocksUnloadEvent
-import io.github.pylonmc.pylon.core.persistence.blockstorage.BlockStorage
 import io.github.pylonmc.pylon.core.util.position.BlockPosition
 import io.github.pylonmc.pylon.core.util.position.ChunkPosition
 import io.github.pylonmc.pylon.core.util.position.position
