+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/block/PhantomBlock.kt
@@ -1,9 +1,5 @@
-package io.github.pylonmc.pylon.core.persistence.blockstorage
+package io.github.pylonmc.pylon.core.block
 
-import io.github.pylonmc.pylon.core.block.BlockCreateContext
-import io.github.pylonmc.pylon.core.block.BlockItemReason
-import io.github.pylonmc.pylon.core.block.PylonBlock
-import io.github.pylonmc.pylon.core.block.PylonBlockSchema
 import io.github.pylonmc.pylon.core.item.ItemStackBuilder
 import io.github.pylonmc.pylon.core.pluginInstance
 import io.papermc.paper.datacomponent.DataComponentTypes
