+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/item/base/BlockPlacer.kt
@@ -3,7 +3,7 @@ package io.github.pylonmc.pylon.core.item.base
 import io.github.pylonmc.pylon.core.block.BlockCreateContext
 import io.github.pylonmc.pylon.core.block.PylonBlock
 import io.github.pylonmc.pylon.core.block.PylonBlockSchema
-import io.github.pylonmc.pylon.core.persistence.blockstorage.BlockStorage
+import io.github.pylonmc.pylon.core.block.BlockStorage
 import org.bukkit.block.Block
 
 interface BlockPlacer {
