+++ pylon-test/src/main/java/io/github/pylonmc/pylon/test/test/block/BlockStorageChunkReloadTest.java
@@ -2,7 +2,7 @@ package io.github.pylonmc.pylon.test.test.block;
 
 import io.github.pylonmc.pylon.core.block.PylonBlock;
 import io.github.pylonmc.pylon.core.block.PylonBlockSchema;
-import io.github.pylonmc.pylon.core.persistence.blockstorage.BlockStorage;
+import io.github.pylonmc.pylon.core.block.BlockStorage;
 import io.github.pylonmc.pylon.core.util.position.ChunkPosition;
 import io.github.pylonmc.pylon.test.block.BlockWithField;
 import io.github.pylonmc.pylon.test.block.Blocks;
