+++ pylon-test/src/main/java/io/github/pylonmc/pylon/test/test/block/SimpleMultiblockRotatedTest.java
@@ -1,6 +1,6 @@
 package io.github.pylonmc.pylon.test.test.block;
 
-import io.github.pylonmc.pylon.core.persistence.blockstorage.BlockStorage;
+import io.github.pylonmc.pylon.core.block.BlockStorage;
 import io.github.pylonmc.pylon.test.base.AsyncTest;
 import io.github.pylonmc.pylon.test.block.Blocks;
 import io.github.pylonmc.pylon.test.util.TestUtil;
