+++ pylon-test/src/main/java/io/github/pylonmc/pylon/test/test/block/TickingBlockErrorTest.java
@@ -1,7 +1,7 @@
 package io.github.pylonmc.pylon.test.test.block;
 
 import io.github.pylonmc.pylon.core.block.TickManager;
-import io.github.pylonmc.pylon.core.persistence.blockstorage.BlockStorage;
+import io.github.pylonmc.pylon.core.block.BlockStorage;
 import io.github.pylonmc.pylon.core.test.GameTestConfig;
 import io.github.pylonmc.pylon.test.PylonTest;
 import io.github.pylonmc.pylon.test.base.GameTest;
